requirements.txt
python-telegram-bot==21.3
openai==1.35.7
sqlalchemy==2.0.31
python-dotenv==1.0.1
aiohttp==3.9.5
